Database changes: Error handling + dependencies support
Errors are now supported through custom Exception subclasses in order to get a clean and proper source code and error handling.

There is now a support for using dependencies in algorithms in the database (first line of a file, add 'needs TYPE/SPECIFIC.LANG TYPE/SPECIFIC.LANG ....')
This should allow to create and ease the creation of more complex algorithms.